arm/chacha-armv8.pl: harmonize key cleansing and add some commentary.
Cleansing is kind of schizophrenic thing in sense that it's intended
as protection against trusted code. Indeed, all code in the current
context is assumed to be trusted, and OS kernel is trusted to provide
adequate isolation between contexts... On the other hand doing
cleansing right is a daunting task, arguably on the verge to counter-
productive. Indeed, doing it right would incur wiping *all* registers
used for computation, as well as *all* off-load storage, as you can't
make an assumption that caller didn't have secrets in non-volatile
register *you* spill on stack. [Or can you? Why caller would trust
you?] So without answers and consensus it's kind of reduced to a
"feel-good" thing, and cleansing key material [and temporary storage
on stack] felt sufficiently good here...