x86_64/x86_64-xlate.pl: pass .cfi directives even to macosx.
This has a side effect, linker starts issuing "could not create
compact unwind for" warnings[!]. Reference to non-compact info is
generated and stack remains back-traceable. Unwinding procedure
can get confused about top-most subroutine name, but otherwise it
works. It remains to be seen if all this can be resolved, but for
now it's deemed safe to ignore the warnings...