x86_64/x86_64-xlate.pl: auto-generate .cfi_startproc/endproc.
It was triggered by MacOS X' tendency to misrepresent current subroutine
name in stack backtrace. Auto-generation is a side effect of moving
.cfi_startproc prior auto-generated endbr64 in prologue. Previous
commit mentions "could not create compact unwind" linker warnings(*).
It appears that one would be better off with non-compact unwinds,
because they seem to be too frail to count on. I mean it's easy to
trick linker to generate bad unwind, and without a single word of
warning. In other words, it's suggested to ignore the warning.

(*) Documentation suggests that you should ask for these warnings,
    not get them by default. And no, documentation doesn't tell how
    to suppress them. It's possible to omit compact unwinds altogether
    with undocumented -no_compact_unwind.