arm/*-armv8.pl: add support for Morello, CHERI aarch64 processor.
Long-input performance results are virtually indistinguishable from
those for Cortex-A76. [Which is why they are not explicitly listed.]
Short-input performance is naturally lower and appears to be affected
mostly by the additional instructions compiler issues to adjust the
"capabilities," the 'C' in 'CHERI'. The overhead varies depending on
the algorithm and how the benchmark loop is organized, one can note
marginal difference, 20%, 40%, ...